# joy/backend/data/src/main/java/ch/ti8m/paybox/entity/purchase/PurchaseProduct.java
package ch.ti8m.paybox.entity.purchase;

import ch.ti8m.paybox.entity.store.Product;
import com.fasterxml.jackson.annotation.JsonBackReference;
import com.fasterxml.jackson.annotation.JsonManagedReference;

import javax.persistence.*;

/**
 * Created by Fatih on 10.05.16.
 */
@Entity
@Table
public class PurchaseProduct{
    private Purchase purchase;
    private Product product;
    private Long productId;
    private double price;
    private int count;
    private long id;

    @Id
    @GeneratedValue(strategy = GenerationType.AUTO)
    public long getId() {
        return id;
    }

    public void setId(long id) {
        this.id = id;
    }

    @ManyToOne(optional = false, fetch = FetchType.LAZY)
    @JsonBackReference
    public Purchase getPurchase() {
        return purchase;
    }

    public void setPurchase(Purchase purchase) {
        this.purchase = purchase;
    }

    @ManyToOne(optional = false)
    public Product getProduct() {
        return product;
    }

    public void setProduct(Product product) {
        this.product = product;
    }

    @Transient
    public Long getProductId() {
        if (product != null) {
            return product.getId();
        } else {
            return productId;
        }
    }

    public void setProductId(Long productId) {
        this.productId = productId;
    }

    public double getPrice() {
        return price;
    }

    public void setPrice(double price) {
        this.price = price;
    }

    public int getCount() {
        return count;
    }

    public void setCount(int count) {
        this.count = count;
    }


}