# joy/backend/data/src/main/java/ch/ti8m/paybox/entity/store/Beacon.java
package ch.ti8m.paybox.entity.store;


import com.fasterxml.jackson.annotation.JsonBackReference;

import javax.persistence.*;

/**
 * Created by Fatih on 10.05.16.
 */
@Entity
@Table
public class Beacon {
    private Long id;
    private int majorId;
    private int minorId;
    private Shop shop;

    @Id
    @GeneratedValue(strategy = GenerationType.AUTO)
    public Long getId() {
        return id;
    }

    public void setId(Long id) {
        this.id = id;
    }

    public int getMajorId() {
        return majorId;
    }

    public void setMajorId(int majorId) {
        this.majorId = majorId;
    }

    public int getMinorId() {
        return minorId;
    }

    public void setMinorId(int minorId) {
        this.minorId = minorId;
    }

    @OneToOne
    @JsonBackReference
    public Shop getShop() {
        return shop;
    }

    public void setShop(Shop shop) {
        this.shop = shop;
    }
}