# joy/backend/data/src/main/java/ch/ti8m/paybox/entity/store/ShopProduct.java
package ch.ti8m.paybox.entity.store;

import javax.persistence.*;
import java.io.Serializable;

/**
 * Created by Fatih on 10.05.16.
 */
@Entity
@Table
public class ShopProduct implements Serializable{
    private Long id;
    private Shop shop;
    private Product product;
    private double price;
    private String special;

    @Id
    @GeneratedValue(strategy = GenerationType.AUTO)
    public Long getId() { return id; }

    public void setId(Long id) { this.id = id; }

    @ManyToOne
    public Shop getShop() {
        return shop;
    }

    public void setShop(Shop shop) {
        this.shop = shop;
    }

    @ManyToOne(fetch = FetchType.EAGER, cascade = CascadeType.ALL)
    public Product getProduct() {
        return product;
    }

    public void setProduct(Product product) {
        this.product = product;
    }

    public double getPrice() {
        return price;
    }

    public void setPrice(double price) {
        this.price = price;
    }

    public String getSpecial() {
        return special;
    }

    public void setSpecial(String special) {
        this.special = special;
    }
}